+++ groove-auth-api/src/main/java/org/bogus/groove_auth/endpoint/AuthAppControllerAdvice.java
@@ -1,51 +1,21 @@
 package org.bogus.groove_auth.endpoint;
 
-import org.bogus.groove.common.AppException;
-import org.bogus.groove.common.BadRequestException;
+import lombok.RequiredArgsConstructor;
 import org.bogus.groove.common.CommonResponse;
-import org.bogus.groove.common.ForbiddenException;
-import org.bogus.groove.common.InternalServerException;
-import org.bogus.groove.common.NotFoundException;
-import org.bogus.groove.common.UnauthorizedException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
+import org.bogus.groove_auth.endpoint.support.ExceptionTranslator;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-@ControllerAdvice
+@ControllerAdvice(basePackages = "org.bogus.groove_auth")
+@RequiredArgsConstructor
 public class AuthAppControllerAdvice {
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+    private final ExceptionTranslator exceptionTranslator;
 
-    @ExceptionHandler(AppException.class)
-    public ResponseEntity<CommonResponse<Void>> handleAppException(AppException e) {
-        HttpStatus status = HttpStatus.INTERNAL_SERVER_ERROR;
-
-        if (e instanceof NotFoundException) {
-            status = HttpStatus.NOT_FOUND;
-            logger.info("[{}] {}, {}", status.value(), status.getReasonPhrase(), e.getMessage());
-        } else if (e instanceof BadRequestException) {
-            status = HttpStatus.BAD_REQUEST;
-            logger.info("[{}] {}, {}", status.value(), status.getReasonPhrase(), e.getMessage());
-        } else if (e instanceof UnauthorizedException) {
-            status = HttpStatus.UNAUTHORIZED;
-            logger.info("[{}] {}, {}", status.value(), status.getReasonPhrase(), e.getMessage());
-        } else if (e instanceof ForbiddenException) {
-            status = HttpStatus.FORBIDDEN;
-            logger.info("[{}] {}, {}", status.value(), status.getReasonPhrase(), e.getMessage());
-        } else if (e instanceof InternalServerException) {
-            status = HttpStatus.INTERNAL_SERVER_ERROR;
-            logger.error("[{}] {}, {}", status.value(), status.getReasonPhrase(), e.getMessage());
-        }
-
-        return ResponseEntity.status(status.value())
-            .body(
-                CommonResponse.error(
-                    e.getError().code(),
-                    e.getMessage(),
-                    e.getExtraData()
-                )
-            );
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<CommonResponse<Void>> handleException(Exception e) {
+        return ResponseEntity
+            .status(exceptionTranslator.translateToHttpStatus(e))
+            .body(exceptionTranslator.translateToCommonResponse(e));
     }
 }
